requirements.txt
flask-mysql-connector==1.1.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
